예시코드/stages/ch04/backend/requirements.txt
# 가이드 참조: chapter4_hyperclovax.md L654–L660
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
